+++ src/components/Table/types.ts
@@ -1,6 +1,5 @@
 import React, { FunctionComponent, MouseEventHandler } from 'react';
-import { StyledComponentBase } from 'styled-components';
-import { SubcomponentPropsType } from '../commonTypes';
+import { SubcomponentPropsType, StyledSubcomponentType } from '../commonTypes';
 
 export type ExpansionIconProps = {
   isCollapsed: boolean;
@@ -21,10 +20,10 @@ export interface columnTypes {
     footerContent?: string;
     sortFunction?: (item1: any, item2: any) => boolean;
     isGroupLabel?: boolean;
-    cellComponent?: string & StyledComponentBase<any, {}>;
-    rowComponent?: string & StyledComponentBase<any, {}>;
-    headerCellComponent?: string & StyledComponentBase<any, {}>;
-    groupCellComponent?: string & StyledComponentBase<any, {}>;
+    cellComponent?: StyledSubcomponentType;
+    rowComponent?: StyledSubcomponentType;
+    headerCellComponent?: StyledSubcomponentType;
+    groupCellComponent?: StyledSubcomponentType;
   };
 }
 
@@ -40,14 +39,14 @@ export type TableProps = {
   sortGroups?: boolean;
   showFooter?: boolean;
 
-  StyledCell?: string & StyledComponentBase<any, {}>;
-  StyledContainer?: string & StyledComponentBase<any, {}>;
-  StyledGroupLabelRow?: string & StyledComponentBase<any, {}>;
-  StyledHeader?: string & StyledComponentBase<any, {}>;
-  StyledHeaderCell?: string & StyledComponentBase<any, {}>;
+  StyledCell?: StyledSubcomponentType;
+  StyledContainer?: StyledSubcomponentType;
+  StyledGroupLabelRow?: StyledSubcomponentType;
+  StyledHeader?: StyledSubcomponentType;
+  StyledHeaderCell?: StyledSubcomponentType;
+  StyledRow?: StyledSubcomponentType;
   StyledFooter?: string & StyledComponentBase<any, {}>;
   StyledFooterCell?: string & StyledComponentBase<any, {}>;
-  StyledRow?: string & StyledComponentBase<any, {}>;
 
   cellProps?: SubcomponentPropsType;
   containerProps?: SubcomponentPropsType;